DJANGO/OrchestrationRouteurCISCO/requirements.txt
Django > 4.2.17
netmiko == 4.5.0
ncclient>=0.6.14
mysqlclient==2.1.1
django-cors-headers
ncclient
celery
redis
django_extensions
python-dotenv
pymysql